New upgrade-selection UX (#255)
## Links

- [Jira](https://rewiringamerica.atlassian.net/browse/RAT-689)

## Description

This puts the upgrade selection onto the same page as the form, and
presents it as radio buttons instead of as a series of submit buttons.

This is a significant refactor of `RemForm`, `UpgradeOptions`, and the
top-level element. All the form state is stored in the top-level
element and passed down to `RemForm` and `UpgradeOptions`, making the
latter two essentially controlled components. This was necessary
because neither of them has a way to "submit" anymore, and thus no
other way to pass form state up to the top-level element.

Upon submitting the form, the top-level element converts the values
into labels for the form snapshot element and stores the labels in
state, and that's what signals it to display the "snapshot + (spinner
| results)" view.

I first tried using Headless UI's `RadioGroup` in UpgradeOptions, but
unfortunately when testing its keyboard navigation, I found it has the
same problem I ran into long ago with `Popover`: it doesn't work right
inside a shadow DOM. When you try to navigate the radio group with Tab
or arrow keys, it ends up looking for focusable elements in the main
document, not in the shadow DOM. So I basically rolled my own :/

## Test Plan

Fill out and submit the form; make sure results are shown as
before. Click the "Edit" button to make sure you're taken back to the
initial page.

Choose a water heating fuel, select water heater as the upgrade, then
choose "no selection" for WH fuel. Make sure the upgrade selection is
cleared, and the submit button is disabled until another upgrade is
clicked.

Navigate the form with a keyboard. It works acceptably, although I
think it's not gold-standard level (e.g. you can't navigate the radio
buttons with arrow keys).