Minor BIC design tweaks
## Links

- [Jira](https://rewiringamerica.atlassian.net/browse/RAT-719)

## Description

- The non-full-width separators was a design update I missed when I
  implemented the "all inputs on one page" flow.

- The main button lightened when clicked, even while disabled. Fixed that.

## Test Plan

Look at the separator in the form, and in the snapshot.

While the submit button is disabled, hover over it and click it; make
sure there are no visual changes for either event. When it's enabled,
hover it and make sure it darkens; click it and make sure it lightens
while the mouse button is down.