Proper error flow for BIC (#259)
## Links

- [Jira](https://rewiringamerica.atlassian.net/browse/RAT-652)

## Description

Read the machine-friendly `type` from the REM API's error response,
and use that to display error text under the address field. (All the
current possible errors are because of the address, or because of a
combination of inputs including the address.)

## Test Plan

Trigger errors with a variety of inputs:

- Address not parsable: `asdf asdf asdf asdf`
- Building not supported: `4350 sw 14th st, miami, fl` and fuel oil for
heating
- Building type not supported: `800 boylston st, boston, ma`
- Country not supported: `20 constance st, toronto, on`
- Multifamily not supported: `253 w 73rd st, new york, ny`
- State not supported: `1017 alewa dr, honolulu, hi`

"Third party error" and the others are not triggerable with user
inputs alone.

Run some real inputs (single-family residential addresses with
realistic fuel choices) and make sure those still work.