More detailed & earlier "ending soon" chips for IRA incentives (#262)
## Links

- [Jira](https://rewiringamerica.atlassian.net/browse/RAT-692)

## Description

We want to prominently highlight the imminent end of the IRA tax
credits, so start showing these chips on federal incentives (all of
which are currently IRA tax credits) 6 months in advance instead of 2,
and include the actual end date in the warning chip.

For non-federal credits, stay at 60 days in advance, but do include
the date in the chip if an exact day is known.

Also I just want to say that Javascript's `Date` is wretched garbage
and I hate it so much argghgh

## Test Plan

Look at the heat pump IRA credits under the HVAC project; make sure
the chip is there and has the right date. Put the embed in Spanish
mode and make sure the chip and date are localized.

Look at the EV charger credit under "transportation" and make sure it
has no chip.