Don't add the stylesheet with appendChild
## Links

- [Jira](https://rewiringamerica.atlassian.net/browse/RAT-679)

## Description

We have a potential embedder who wants to embed the calculator on a
page with a locked-down Content-Security-Policy. However, the
calculator loads its stylesheet by creating a `<style>` DOM node in JS
and appending it to the DOM tree, and the only way that can pass a CSP
is with an `unsafe-inline` directive, which defeats the purpose.

This switches to using a `<link>` tag to load the stylesheet, which
can pass a `style-src` directive in a CSP.

The change to `bill-impact-calculator.html` to remove the `<link>` to
the stylesheet is not incidental; it's required to make this work. I
believe this is a quirk (bug?) of Parcel, but if I keep that link
there, the code that loads the stylesheets in the calculator elements
fails in dev mode, with an "invalid URL" error.

## Test Plan

Look at the preview deploy of this branch and make sure it has the
right styles and works correctly.

I pushed another branch (`owen.csp-test2`) in which I put a `<meta
http-equiv="Content-Security-Policy" ... />` tag on `index.html`, with
the directives I put in the README.

First I changed index.html on that branch to load
`state-calculator.js` and `rewiring-fonts.css` from
`https://embed.rewiringamerica.org` (instead of using relative
paths). I made sure the calculator failed to load due to CSP errors,
and confirmed that the `appendChild` of the stylesheet was the failing
line of code.

Then I changed index.html to load the JS and CSS files from this PR's
branch deploy, updated the domains in the meta tag's CSP directive
accordingly, and made sure the calculator loads, looks correct, and
functions (including making API calls successfully).