Simplify Parcel config slightly (#285)
## Description

The CSP issue stems from some weirdness in the way Parcel works
(possibly a bug?) and I came across this while debugging. This change:

- Uses a newer alternative to the transformer-svg-react plugin, that
does essentially the same thing. [This is what the docs
recommend](https://parceljs.org/languages/svg/#importing-as-a-react-component).

- Simplifies the config to avoid explicitly listing out the
  transformers for other file types, and instead fall back to what the
  default Parcel config specifies. This is better, in case the default
  config ever changes.

Seems like simpler is better in any case.

(I checked: this by itself does not solve the issue with the CSP thing
I'm doing.)

## Test Plan

- `yarn tsc --noEmit` passes.

- In both dev mode (`yarn serve:widget`) and prod mode
  (`yarn build:widget && npx serve dist`), do:

  - In the incentives embed, submit the form.

  - Make sure the project graphic shows up before you choose a project.

  - Make sure the icons show up in the project selector. The one shown
    in the dropdown when it's closed should be purple; in the dropdown
    menu, they should be black.