Second attempt at using <link> to load CSS (#286)
## Links

- [Jira](https://rewiringamerica.atlassian.net/browse/RAT-679)
- [Parcel bug
report](https://github.com/parcel-bundler/parcel/issues/10213)

## Description

This is a repeat of #281 (which was since reverted as #283), but with
a workaround for a Parcel bug that it triggered. (I've filed a bug
report on Parcel, linked above.)

The problem with the original PR was that Parcel turned the
`new URL('../tailwind.css', import.meta.url)` calls into an invalid URL
**if** the `<script>` tag that loaded the embed JS appeared on the
page **after** the embed element itself. This meant that the
calculator UI components all showed up and were fully functional, but
they were unstyled.

For whatever reason, this bug goes away if I remove
`src/calculator.ts` from the list of files for Parcel to build. So I
did that. To continue supporting embedders who are loading the embed
JS from `/calculator.js` and using the custom tag
`rewiring-america-calculator`, the file `state-calculator.js` now
defines both custom tags, and we manually copy that file to
`calculator.js` immediately after running `parcel build`.

To make sure we're poking this bug during development, I updated the
`serve:widget` command to be as close as possible to the prod build
command, and moved the script tags on the demo pages to be after the
custom components.

This whole situation feels upsettingly brittle, so I'll keep an eye on
that Parcel bug report and possibly see if I can fix it myself.

## Test Plan

(Try just reverting #283 and doing the below steps to repro the bug)

- Run `yarn build:widget` and `npx serve dist`. This serves the built
  artifacts on port 3000.

- In a separate directory, create an HTML page with the
  state-calculator custom component on it, put the `<script>` tag
  below it, loading from `http://localhost:3000/state-calculator.js`.

- Run `npx serve` in that separate directory, then go to the URL it
  prints in a browser. Make sure the calculator shows up and is
  properly styled.

- Change the script tag to load `http://localhost:3000/calculator.js`,
  change the HTML tag to `rewiring-america-calculator`, and make sure
  that still brings up the calculator, correctly styled.

- Change the script tag to load
  `http://localhost:3000/bill-impact-calculator.js`, change the HTML
  tag to `rewiring-america-bill-impact-calculator`, and make sure that
  brings up BIC, correctly styled (and the logo appears at the top).

Then repeat the test plan from #281 to test the CSP part.